cleaned a few dirty statements found by g95 (but not compiling yet)

+++ input/import_h2_dalton.f
@@ -27,7 +27,7 @@
       integer, parameter ::
      &     ntest = 00
 
-      integer, parameter ::
+      integer(4), parameter ::
      &     imsk16 = 65535,
      &     imsk08 =   255
